+++ src/component/Legend.tsx
@@ -1,10 +1,11 @@
-import React, { PureComponent, CSSProperties, useState, useCallback } from 'react';
+import React, { CSSProperties, PureComponent } from 'react';
 import { DefaultLegendContent, Payload, Props as DefaultProps } from './DefaultLegendContent';
 
 import { isNumber } from '../util/DataUtils';
 import { LayoutType } from '../util/types';
-import { UniqueOption, getUniqPayload } from '../util/payload/getUniqPayload';
+import { getUniqPayload, UniqueOption } from '../util/payload/getUniqPayload';
 import { useLegendPayload } from '../context/legendPayloadContext';
+import { BoundingBox, useGetBoundingClientRect } from '../util/useGetBoundingClientRect';
 
 function defaultUniqBy(entry: Payload) {
   return entry.value;
@@ -33,15 +34,6 @@ function LegendContent(props: ContentProps) {
 
   return <DefaultLegendContent {...propsWithoutRef} />;
 }
-
-const EPS = 1;
-
-type BoundingBox = {
-  width: number;
-  height: number;
-};
-
-type SetBoundingBox = (node: HTMLElement | null) => void;
 
 function getDefaultPosition(style: React.CSSProperties, props: Props, box: BoundingBox) {
   const { layout, align, verticalAlign, margin, chartWidth, chartHeight } = props;
@@ -92,43 +84,6 @@ export type Props = DefaultProps & {
 interface State {
   boxWidth: number;
   boxHeight: number;
-}
-
-/**
- * Use this to listen to bounding box changes.
- * https://developer.mozilla.org/en-US/docs/Web/API/Element/getBoundingClientRect
- *
- * Very useful for reading actual sizes of DOM elements.
- *
- * Be aware that this is difficult to test for. jsdom, by design, returns all zeroes!
- * If you want to write tests check out the utility function `mockGetBoundingClientRect`
- *
- * @param onUpdate this is extra callback that's called with the full DOMRect. Note that this receives different object than the return value.
- * @param extraDependencies use this to trigger new DOM dimensions read when any of these change. Good for things like payload and label, that will re-render something down in the children array, but you want to read the bounding box of a parent.
- * @returns [lastBoundingBox, updateBoundingBox] most recent value, and setter. Pass the setter to a DOM element ref like this: `<div ref={updateBoundingBox}>`
- */
-function useGetBoundingClientRect(
-  onUpdate: (domRect: DOMRect | null) => void,
-  extraDependencies: ReadonlyArray<unknown>,
-): [BoundingBox, SetBoundingBox] {
-  const [lastBoundingBox, setLastBoundingBox] = useState<BoundingBox>({ width: 0, height: 0 });
-  const updateBoundingBox = useCallback(
-    (node: HTMLDivElement | null) => {
-      if (node != null && node.getBoundingClientRect) {
-        const box = node.getBoundingClientRect();
-        box.width = node.offsetWidth;
-        box.height = node.offsetHeight;
-        if (Math.abs(box.width - lastBoundingBox.width) > EPS || Math.abs(box.height - lastBoundingBox.height) > EPS) {
-          setLastBoundingBox({ width: box.width, height: box.height });
-          if (onUpdate) {
-            onUpdate(box);
-          }
-        }
-      }
-    },
-    [lastBoundingBox.width, lastBoundingBox.height, onUpdate, ...extraDependencies],
-  );
-  return [lastBoundingBox, updateBoundingBox];
 }
 
 function LegendWrapper(props: Props) {
